// frontend/client/src/types.tsx
export type Permissions = {
  [key in typeof RESOURCES[number]]?: typeof PERMISSIONS[number][];
};
export type APIError = { status: number; moreInfo: string; message: string };
export type ScheduledBroadcastResponse = { results: ScheduledBroadcast[] };
export type BroadcastResponse = { results: Broadcast[] };
export type UserResponse = { results: User };
// ...
export type User = {
  uuid?: string;
  id: number;
  email: string;
  username: string;
  permissions: Permissions;
};